Fix issues with the PCRE library and speedup build
This PR fixes two issues at once (I know it's not the best practice,
but these changes are tied together, and working on them in
multiple branches will lead to merge conflicts):

1. The "grave" issue: "Compilation failed: unrecognised compile-time
option bit(s) at offset 0".  This one seems to be caused by an
[outdated PCRE library](https://www.mail-archive.com/debian-bugs-dist@lists.debian.org/msg1830385.html);
the library has to be updated from the PPA the PHP is installed from.
2. Build failures resulting from transient network consitions (like
"Unable to connect to archive.ubuntu.com")
3. Speed up the build process by using [eatmydata](https://manpages.debian.org/testing/eatmydata/eatmydata.1.en.html)